requirements.txt
aiosqlite==0.20.0
playwright==1.49.0
playwright-stealth==1.0.6
httpx==0.28.1
tenacity==9.0.0
rapidfuzz==3.11.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
python-telegram-bot==21.8
pytesseract==0.3.13
pillow==11.0.0
python-dotenv==1.0.1
